Implement Brave Ads Serving Improvements
Currently, ad serving is done by randomly picking an ad from matching segments. Problems are under-delivery of creatives and users not seeing as many ads as they could. This PR aims to solve these short comings by changing the way we serve ads. We introduce a statistical model for ad selection and assume that the click probability can be approximated by a linear function of features related to locally observed user behaviour. The function is weighted by a control-parameter called pass-through rate (ptr) to throttle delivery of individual ads.